requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiodns==4.0.0
    # via ccxt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   ccxt
    #   python-binance
aiosignal==1.4.0
    # via aiohttp
attrs==25.4.0
    # via aiohttp
bleach==6.3.0
    # via panel
bokeh==3.8.2
    # via
    #   -r requirements.in
    #   dask
    #   holoviews
    #   panel
build==1.4.0
    # via pip-tools
ccxt==4.5.40
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   ccxt
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   pycares
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   dask
    #   distributed
    #   pip-tools
cloudpickle==3.1.2
    # via
    #   dask
    #   distributed
coincurve==21.0.0
    # via ccxt
colorcet==3.1.0
    # via holoviews
contourpy==1.3.3
    # via bokeh
cryptography==46.0.5
    # via
    #   -r requirements.in
    #   ccxt
dask[array,complete,dataframe,diagnostics,distributed]==2026.1.2
    # via
    #   -r requirements.in
    #   distributed
dateparser==1.3.0
    # via python-binance
distributed==2026.1.2
    # via dask
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.2.0
    # via dask
holoviews==1.22.1
    # via -r requirements.in
idna==3.11
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   bokeh
    #   dask
    #   distributed
joblib==1.5.3
    # via scikit-learn
librt==0.8.1
    # via mypy
linkify-it-py==2.0.3
    # via panel
llvmlite==0.43.0
    # via numba
locket==1.0.0
    # via
    #   distributed
    #   partd
lz4==4.4.5
    # via dask
markdown==3.10.2
    # via panel
markdown-it-py==4.0.0
    # via
    #   mdit-py-plugins
    #   panel
markupsafe==3.0.3
    # via jinja2
mdit-py-plugins==0.5.0
    # via panel
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via
    #   -r requirements.in
    #   distributed
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy==1.19.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
narwhals==2.17.0
    # via
    #   bokeh
    #   holoviews
    #   panel
numba==0.60.0
    # via -r requirements.in
numpy==2.0.2
    # via
    #   -r requirements.in
    #   bokeh
    #   contourpy
    #   dask
    #   holoviews
    #   numba
    #   pandas
    #   pandas-stubs
    #   scikit-learn
    #   scipy
packaging==26.0
    # via
    #   bokeh
    #   build
    #   dask
    #   distributed
    #   holoviews
    #   panel
    #   wheel
pandas==3.0.1
    # via
    #   -r requirements.in
    #   bokeh
    #   dask
    #   holoviews
    #   panel
pandas-stubs==3.0.0.260204
    # via -r requirements.in
panel==1.8.7
    # via holoviews
param==2.3.2
    # via
    #   holoviews
    #   panel
    #   pyviz-comms
partd==1.4.2
    # via dask
pathspec==1.0.4
    # via mypy
pillow==12.1.1
    # via bokeh
pip-tools==7.5.3
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psutil==7.2.2
    # via distributed
pyarrow==23.0.1
    # via dask
pycares==5.0.1
    # via aiodns
pycparser==3.0
    # via cffi
pycryptodome==3.23.0
    # via
    #   -r requirements.in
    #   python-binance
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-binance==1.0.35
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   holoviews
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via dateparser
pyviz-comms==3.0.6
    # via
    #   holoviews
    #   panel
pyyaml==6.0.3
    # via
    #   bokeh
    #   dask
    #   distributed
regex==2026.2.19
    # via dateparser
requests==2.32.5
    # via
    #   ccxt
    #   panel
    #   python-binance
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.17.1
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.17.0
    # via
    #   python-binance
    #   python-dateutil
sortedcontainers==2.4.0
    # via
    #   -r requirements.in
    #   distributed
tblib==3.2.2
    # via distributed
threadpoolctl==3.6.0
    # via scikit-learn
toolz==1.1.0
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.5.4
    # via
    #   bokeh
    #   distributed
tqdm==4.67.3
    # via panel
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   ccxt
    #   mypy
    #   panel
tzlocal==5.3.1
    # via dateparser
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.6.3
    # via
    #   distributed
    #   requests
webencodings==0.5.1
    # via bleach
websockets==16.0
    # via python-binance
wheel==0.46.3
    # via pip-tools
xyzservices==2025.11.0
    # via bokeh
yarl==1.22.0
    # via
    #   aiohttp
    #   ccxt
zict==3.0.0
    # via distributed

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
